Add command based group setup disable signal approach.
